+++ airflow/dags/initial_database_setup_dag.py
@@ -7,7 +7,7 @@ from datetime import datetime
 import sys
 sys.path.append('/opt/workspace/facial_database/python_scripts/')
 
-import initial_database_setup as db
+import setup_initial_database as db
 
 ###############################################
 # Parameters
