+++ facial_database/python_scripts/ops_download_files.py
@@ -1,4 +1,4 @@
-import check_packages as cp
+import ops_check_packages as cp
 
 packages_required = ["pandas","requests"]
